+++ src/dr/evomodel/operators/SubtreeJumpOperator.java
@@ -135,6 +135,10 @@ public class SubtreeJumpOperator extends AbstractTreeOperator implements Constra
         final NodeRef j = destinations.get(r);
         final NodeRef jP = tree.getParent(j);
 
+        // the distance travelled in branch length space - the distance between node i and j minus how much
+        // higher i is than node j.
+        double delta = TreeUtils.getPathLength(tree, i, j) - (tree.getNodeHeight(i) - tree.getNodeHeight(j));
+
         tree.beginTreeEdit();
 
         // remove the parent of i by connecting its sibling to its grandparent.
@@ -163,6 +167,8 @@ public class SubtreeJumpOperator extends AbstractTreeOperator implements Constra
             // hastings ratio = reverse Prob / forward Prob
             logq = Math.log(reverseProbability) - Math.log(forwardProbability);
         }
+
+        setTopologyChanged(true, delta);
 
         return logq;
     }