+++ src/dr/evomodelxml/speciation/SpeciationLikelihoodParser.java
@@ -32,7 +32,8 @@ import dr.evolution.util.Taxa;
 import dr.evolution.util.Taxon;
 import dr.evolution.util.TaxonList;
 import dr.evomodel.speciation.CalibrationPoints;
-import dr.evomodel.birthdeath.EfficientBirthDeathLikelihood;
+import dr.evomodel.birthdeath.BirthDeathModel;
+import dr.evomodel.speciation.EfficientBirthDeathSpeciationLikelihood;
 import dr.evomodel.speciation.SpeciationLikelihood;
 import dr.evomodel.speciation.SpeciationModel;
 import dr.inference.distribution.DistributionLikelihood;
@@ -191,6 +192,11 @@ public class SpeciationLikelihoodParser extends AbstractXMLObjectParser {
             } catch( IllegalArgumentException e ) {
                 throw new XMLParseException( e.getMessage() );
             }
+        }
+
+        boolean useNewLoop = xo.getAttribute(USE_NEW_LOOP, false);
+        if (useNewLoop && specModel instanceof BirthDeathModel) {
+            return new EfficientBirthDeathSpeciationLikelihood(tree, (BirthDeathModel) specModel, excludeTaxa, null);
         }
 
         return new SpeciationLikelihood(tree, specModel, excludeTaxa, null);