+++ src/dr/inference/markovchain/MarkovChain.java
@@ -50,7 +50,7 @@ public final class MarkovChain implements Serializable {
     private static final long serialVersionUID = 181L;
 
     private final static boolean DEBUG = false;
-    private final static boolean PROFILE = true;
+    private final static boolean PROFILE = false;
 
     public static final double EVALUATION_TEST_THRESHOLD = 1e-1;